Update Xcode project to use the Swift package as a local dependency
SwiftPM is the supported method of creating a macro plugin. Remove the
Xcode project's `Other Swift Flags` and build the plugin through the
package instead. This is a little weird as we have `MacroExamples` as a
target in both the project and the package - we could remove the project
entirely, but it seems nice to still have it with the package dependency
as an example.